Update various important things in preparation for first full run
Fix bug in generating input/target CSV file for Operon, where an extra
newline at the end of the file was erroneously left out previously.
(Operon seems to expect this internally.) Configure DEAP profiling code
to store fitness outputs, rather than the relevant `evaluate.py` script,
since DEAP seems to be faster for whatever reason. Update all scripts
to be in the context of an initial full run for 32 bins of 512 programs.